+++ app/models/bloom_trade_client/conversion_result.rb
@@ -1,10 +1,10 @@
 module BloomTradeClient
   class ConversionResult
-    attr_reader :rate, :expires_at
 
-    def initialize(rate, expires_at="")
-      @rate = rate
-      @expires_at = expires_at
-    end
+    include Virtus.model
+
+    attribute :rate, Float
+    attribute :expires_at, Integer, default: 0
+
   end
 end