+++ app/services/bloom_trade_client/exchange_rates/convert.rb
@@ -14,12 +14,12 @@ module BloomTradeClient
         origin_rate = direct_rate(type, base_currency, reserve_currency)
         destination_rate = direct_rate(type, reserve_currency, counter_currency)
         reverse_rate = origin_rate.rate * destination_rate.rate if origin_rate && destination_rate
-        return ConversionResult.new(reverse_rate) if reverse_rate
+        return ConversionResult.new(rate: reverse_rate) if reverse_rate
 
         Rails.logger.error(
           "Unable to calculate rate #{base_currency}#{counter_currency}"
         )
-        return ConversionResult.new(0.0)
+        return ConversionResult.new(rate: 0.0)
       end
 
       private
@@ -27,7 +27,7 @@ module BloomTradeClient
       def self.direct_rate(type, base_currency, counter_currency)
         return nil unless %w(buy sell mid).include? type
 
-        return ConversionResult.new(1.0) if base_currency == counter_currency
+        return ConversionResult.new(rate: 1.0) if base_currency == counter_currency
 
         exchange_rate = ExchangeRate.find_by(
           base_currency: base_currency,
@@ -36,8 +36,8 @@ module BloomTradeClient
 
         if exchange_rate
           return ConversionResult.new(
-            exchange_rate.send(type.to_sym),
-            exchange_rate.expires_at
+            rate: exchange_rate.send(type.to_sym),
+            expires_at: exchange_rate.expires_at
           )
         end
 
@@ -48,8 +48,8 @@ module BloomTradeClient
 
         if reversed_rate
           return ConversionResult.new(
-            1.0 / reversed_rate.send(type.to_sym),
-            reversed_rate.expires_at
+            rate: 1.0 / reversed_rate.send(type.to_sym),
+            expires_at: reversed_rate.expires_at
           )
         end
       end